+++ hazelcast/include/hazelcast/client/IMap.h
@@ -23,7 +23,6 @@
 #include <stdexcept>
 #include <climits>
 #include <hazelcast/client/protocol/codec/MapAddEntryListenerWithPredicateCodec.h>
-#include "hazelcast/client/query/PagingPredicate.h"
 #include "hazelcast/client/proxy/IMapImpl.h"
 #include "hazelcast/client/impl/EntryEventHandler.h"
 #include "hazelcast/client/EntryListener.h"
@@ -635,7 +634,7 @@ namespace hazelcast {
                     entries[i] = std::make_pair(*key, V());
                 }
 
-                std::pair<int, int> range = getSortedQueryResultSet<Compare>(entries, predicate, query::KEY);
+                std::pair<int, int> range = getSortedQueryResultSet<K, V, Compare>(entries, predicate, query::KEY);
 
                 // return the sublist of entries
                 std::vector<K> result;
@@ -726,7 +725,7 @@ namespace hazelcast {
                     entries[i] = std::make_pair(*key, *value);
                 }
 
-                std::pair<int, int> range = getSortedQueryResultSet<Compare>(entries, predicate, query::VALUE);
+                std::pair<int, int> range = getSortedQueryResultSet<K, V, Compare>(entries, predicate, query::VALUE);
 
                 // return the sublist of entries
                 std::vector<V> result;
@@ -826,7 +825,7 @@ namespace hazelcast {
                     entries[i] = std::make_pair(*key, *value);
                 }
 
-                std::pair<int, int> range = getSortedQueryResultSet<Compare>(entries, predicate, query::ENTRY);
+                std::pair<int, int> range = getSortedQueryResultSet<K, V, Compare>(entries, predicate, query::ENTRY);
                 typename std::vector<std::pair<K, V> >::iterator start = entries.begin();
                 return std::vector<std::pair<K, V> >(start + range.first, start + range.second);
             }
@@ -1019,71 +1018,6 @@ namespace hazelcast {
         private:
             IMap(const std::string &instanceName, spi::ClientContext *context)
                     : proxy::IMapImpl(instanceName, context) {
-            }
-
-            template <typename Compare>
-            std::pair<int, int> getSortedQueryResultSet(std::vector<std::pair<K, V> > &entries, query::PagingPredicate<K, V, Compare> &predicate, query::IterationType iterationType) {
-                if (entries.empty()) {
-                    return std::pair<int, int>(0, 0);
-                }
-
-                sort<Compare>(entries, predicate, iterationType);
-
-                const std::pair<int, std::pair<K, V> > *nearestAnchorEntry = predicate.getNearestAnchorEntry();
-                int nearestPage = (NULL == nearestAnchorEntry ? -1 : nearestAnchorEntry->first);
-                int page = predicate.getPage();
-                int pageSize = predicate.getPageSize();
-                int begin = pageSize * (page - nearestPage - 1);
-                int size = (int)entries.size();
-                if (begin > size) {
-                    return std::pair<int, int>(0, 0);
-                }
-                int end = begin + pageSize;
-                if (end > size) {
-                    end = size;
-                }
-
-                setAnchor(entries, predicate, nearestPage);
-
-                return std::pair<int, int>(begin, end);
-            }
-
-            template <typename Compare>
-            void sort(std::vector<std::pair<K, V> > &entries, const query::PagingPredicate<K, V, Compare> &predicate, query::IterationType iterationType) const {
-                if (NULL != predicate.getComparator()) {
-                    std::sort<typename std::vector<std::pair<K, V> >::iterator, Compare>(entries.begin(), entries.end(), *predicate.getComparator());
-                } else {
-                    switch (iterationType) {
-                        case query::VALUE:
-                        {
-                            query::ValueComparator<K, V> comp;
-                            std::sort<typename std::vector<std::pair<K, V> >::iterator, query::ValueComparator<K, V> >(entries.begin(), entries.end(), comp);
-                            break;
-                        }
-                        case query::KEY:
-                        case query::ENTRY:
-                        {
-                            query::KeyComparator<K, V> comp;
-                            std::sort<typename std::vector<std::pair<K, V> >::iterator, query::KeyComparator<K, V> >(entries.begin(), entries.end(), comp);
-                            break;
-                        }
-                    }
-                }
-            }
-
-            template <class Compare>
-            static void setAnchor(std::vector<std::pair<K, V> > &entries, query::PagingPredicate<K, V, Compare> &predicate, int nearestPage) {
-                if (entries.empty()) {
-                    return;
-                }
-
-                size_t size = entries.size();
-                size_t pageSize = (size_t)predicate.getPageSize();
-                for (size_t i = pageSize; i <= size; i += pageSize) {
-                    const std::pair<K, V> &anchor = entries[i - 1];
-                    nearestPage++;
-                    predicate.setAnchor(nearestPage, anchor);
-                }
             }
         };
     }
